Moved logic to Recipe class
